Move ToggleProvider to theme
